requirements.txt
asgiref==3.8.1
certifi==2025.4.26
crispy-tailwind==1.0.3
Django==5.2.1
django-crispy-forms==2.4
django-htmx==1.23.0
django-meta==2.5.0
django-robots==6.1
pillow==11.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
sib-api-v3-sdk==7.6.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
